refactor(GenericDropdownStyles): new styles for dropdown when the placeholder is selected
